+++ headers/math32.h
@@ -4,7 +4,7 @@
 #include "types.h"
 
 i32 abs(i32);
-i32 sign(i32);
+i8 sign(i32);
 u32 udiv32(u32, u32);
 u32 urem32(u32, u32);
 i32 div32(i32, i32);
